requirements.txt
googletrans==4.0.0rc1
keyboard==0.13.5
Pillow==11.0.0
PyAutoGUI==0.9.54
pyperclip==1.9.0
pystray==0.19.5
Windows_Toasts==1.3.0
